Conllu async data
The data comes from the service subscribing to sentence obs
When the cmpt is destroyed unsub is done
